Add(seungbum/board): 게시글 생성, 삭제 Swagger 적용 및 AuthGuard 적용 #11

+++ mohae/src/boards/boards.service.ts
@@ -19,8 +19,9 @@ import { UserRepository } from 'src/auth/repository/user.repository';
 import { Board } from './entity/board.entity';
 import { Category } from 'src/categories/entity/category.entity';
 import { Area } from 'src/areas/entity/areas.entity';
-import { User } from 'src/auth/entity/user.entity';
+// import { UserInfo } from 'src/auth/entity/user.entity';
 import { BoardPhotoRepository } from 'src/photo/repository/photo.repository';
+import { User } from '@sentry/node';
 
 @Injectable()
 export class BoardsService {
@@ -205,13 +206,16 @@ export class BoardsService {
     return { foundedBoardNum: boards.length, search: title, boards };
   }
 
-  async createBoard(createBoardDto: CreateBoardDto): Promise<boolean> {
+  async createBoard(
+    createBoardDto: CreateBoardDto,
+    user: User,
+  ): Promise<boolean> {
     const queryRunner = this.connection.createQueryRunner();
 
     await queryRunner.connect();
     await queryRunner.startTransaction();
     try {
-      const { categoryNo, areaNo, deadline, userNo, photoUrl }: CreateBoardDto =
+      const { categoryNo, areaNo, deadline, photoUrl }: CreateBoardDto =
         createBoardDto;
 
       const category: Category = await this.categoryRepository.findOne(
@@ -231,12 +235,6 @@ export class BoardsService {
       });
 
       this.errorConfirm.notFoundError(area, `해당 지역을 찾을 수 없습니다.`);
-
-      const user: User = await this.userRepository.findOne(userNo, {
-        relations: ['boards'],
-      });
-
-      this.errorConfirm.notFoundError(user, `해당 회원을 찾을 수 없습니다.`);
 
       let endTime: Date = new Date();
 
@@ -289,25 +287,34 @@ export class BoardsService {
     }
   }
 
-  async deleteBoard(boardNo: number): Promise<boolean> {
-    const board: Board = await this.boardRepository.findOne(boardNo);
-    this.errorConfirm.notFoundError(
-      board.no,
-      `해당 게시글을 찾을 수 없습니다.`,
-    );
+  async deleteBoard(boardNo: number, userNo: number): Promise<boolean> {
+    try {
+      const board: Board = await this.boardRepository.getByOneBoard(boardNo);
+      this.errorConfirm.notFoundError(
+        board.no,
+        `해당 게시글을 찾을 수 없습니다.`,
+      );
 
-    const result: number = await this.boardRepository.deleteBoard(boardNo);
+      if (board['userNo'] !== userNo) {
+        throw new BadRequestException('게시글을 작성한 유저가 아닙니다.');
+      }
 
-    if (!result) {
-      throw new BadGatewayException('해당 게시글이 삭제되지 않았습니다.');
+      const result: number = await this.boardRepository.deleteBoard(boardNo);
+
+      if (!result) {
+        throw new BadGatewayException('해당 게시글이 삭제되지 않았습니다.');
+      }
+
+      return true;
+    } catch (err) {
+      throw err;
     }
-
-    return true;
   }
 
   async updateBoard(
     boardNo: number,
     updateBoardDto: UpdateBoardDto,
+    userNo: number,
   ): Promise<boolean> {
     const queryRunner = this.connection.createQueryRunner();
 
@@ -317,8 +324,12 @@ export class BoardsService {
       const { category, area, deadline, photoUrl }: UpdateBoardDto =
         updateBoardDto;
 
-      const board: Board = await this.boardRepository.findOne(boardNo);
+      const board: Board = await this.boardRepository.getByOneBoard(boardNo);
       this.errorConfirm.notFoundError(board, `해당 게시글을 찾을 수 없습니다.`);
+
+      if (board['userNo'] !== userNo) {
+        throw new BadRequestException('게시글을 작성한 유저가 아닙니다.');
+      }
 
       let endTime: Date = new Date(board.createdAt);
